Fixed: unit tests necessarily running right version of Python.
Without this fix, the tests which actually execute pyle.py would do so using whatever `python` was on the path. But the tests might have been run with a special interpreter, like that from inside of a virtualenv, without the path being modified. Now the tests would be run with an interpreter not meant to be tested.

This fix makes sure to use the same interpreter as the one running the tests.